requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncer==0.0.8
attrs==25.4.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
diskcache==5.6.3
distro==1.9.0
dspy==3.1.3
dspy-ai==3.1.3
fastapi==0.135.1
fastuuid==0.14.0
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
gepa==0.0.26
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jiter==0.13.0
json_repair==0.58.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
litellm==1.82.4
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
numpy==2.4.3
openai==2.29.0
optuna==4.8.0
orjson==3.11.7
packaging==26.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
